+++ back/src/users/users.service.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import dataSource from 'db/data-source';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -10,6 +9,7 @@ import { User } from './entities/user.entity';
 export class UsersService {
   constructor(
     @InjectRepository(User) private usersRepository: Repository<User>,
+    @InjectRepository(Event) private eventsRepository: Repository<Event>,
   ) {}
   async create(createUserDto: CreateUserDto) {
     const newUser = this.usersRepository.create(createUserDto);
@@ -17,25 +17,53 @@ export class UsersService {
   }
 
   async findAll() {
-    const users = this.usersRepository.find({
+    const user = this.usersRepository.find({
       relations: {
         events: true,
+        gifts: true,
+      },
+      // where: {
+      //   userCreatorId: user,
+      // },
+    });
+
+    return await user;
+  }
+
+  async findOneById(id: string) {
+    const user = this.usersRepository.find({
+      relations: {
+        events: true,
+        gifts: true,
+      },
+      where: {
+        id: id,
       },
     });
-    return users;
+    return await user;
   }
 
   async findOne(username: string): Promise<User | undefined> {
     return this.usersRepository.findOneBy({ username });
   }
 
-  async update(username: string, updateUserDto: UpdateUserDto) {
-    const user = await this.findOne(username);
-    return this.usersRepository.save({ ...user, ...updateUserDto });
+  /**
+   * Update, remove работают не корректно
+   */
+
+  async update(userId: string, id: string, updateUserDto: UpdateUserDto) {
+    if (userId === id) {
+      const user = await this.findOneById(id);
+      return this.usersRepository.save({ ...user, ...updateUserDto });
+    }
+    return { statusCode: 403, message: 'Запрещено обновлять чужой Аккаунт' };
   }
 
-  async remove(username: string) {
-    const user = await this.findOne(username);
-    return this.usersRepository.remove(user);
+  async remove(userId: string, id: string) {
+    if (userId === id) {
+      const user = await this.findOne(id);
+      return this.usersRepository.remove(user);
+    }
+    return { statusCode: 403, message: 'Запрещено удалять чужой Аккаунт' };
   }
 }
